Move which.py to lib/which.py
